First stab at refactoring existing functionality
