+++ intake_esm/core.py
@@ -6,12 +6,8 @@ from intake.catalog import Catalog
 from intake.catalog.local import LocalCatalogEntry
 
 from . import config as config
-from .cesm import CESMCollection
-from .cmip import CMIP5Collection, CMIP6Collection
-from .common import _get_built_collections, _open_collection
-from .era5 import ERA5Collection
+from .collection import _get_built_collections, _open_collection
 from .gmet import GMETCollection
-from .mpige import MPIGECollection
 
 
 class ESMMetadataStoreCatalog(Catalog):
@@ -38,12 +34,12 @@ class ESMMetadataStoreCatalog(Catalog):
 
     name = 'esm_metadatastore'
     collection_types = {
-        'cesm': CESMCollection,
-        'cmip5': CMIP5Collection,
-        'cmip6': CMIP6Collection,
-        'mpige': MPIGECollection,
+        # 'cesm': CESMCollection,
+        # 'cmip5': CMIP5Collection,
+        # 'cmip6': CMIP6Collection,
+        # 'mpige': MPIGECollection,
         'gmet': GMETCollection,
-        'era5': ERA5Collection,
+        # 'era5': ERA5Collection,
     }
 
     def __init__(
